requirements.txt
Flask==1.1.1
uwsgi==2.0.18
google-cloud-datastore==1.10.0
